[frontend] Replace custom esbuild based build script by Vite and migrate front to ESM (#14044)
