+++ opencti-platform/Dockerfile
@@ -45,7 +45,7 @@ RUN set -ex; \
 COPY opencti-front /opt/opencti-build/opencti-front
 COPY opencti-graphql/config/schema/opencti.graphql /opt/opencti-build/opencti-graphql/config/schema/opencti.graphql
 
-RUN yarn build:standalone
+RUN yarn build
 
 
 FROM base AS app
@@ -72,13 +72,13 @@ RUN pip3 install --no-cache-dir --requirement ./src/python/requirements.txt \
 COPY --from=graphql-deps-builder /opt/opencti-build/opencti-graphql/node_modules ./node_modules
 COPY --from=graphql-builder /opt/opencti-build/opencti-graphql/build ./build
 COPY --from=graphql-builder /opt/opencti-build/opencti-graphql/static ./static
-COPY --from=front-builder /opt/opencti-build/opencti-front/builder/prod/build ./public
+COPY --from=front-builder /opt/opencti-build/dist ./public
 COPY opencti-graphql/src ./src
 COPY opencti-graphql/config ./config
 COPY opencti-graphql/script ./script
 
 ENV PYTHONUNBUFFERED=1
-ENV NODE_OPTIONS=--max_old_space_size=12288
+ENV NODE_OPTIONS="--max_old_space_size=12288"
 ENV NODE_ENV=production
 
 RUN set -ex; \
