+++ opencti-platform/opencti-front/.gitignore
@@ -1,5 +1,6 @@
 .eslintcache
 config/development.json
+dist/
 
 # testing
 /coverage
