--- opencti-platform/opencti-front/builder/dev/index.html
@@ -1,20 +1,0 @@
-<!doctype html>
-<html lang="en">
-<head>
-    <script>window.BASE_PATH = "%BASE_PATH%"</script>
-    <script defer="defer" src="/js/front.js"></script>
-    %APP_SCRIPT_SNIPPET%
-    <title>%APP_TITLE%</title>
-    <meta http-equiv="Content-Type" content="text/html; charset=utf-8">
-    <meta http-equiv="X-UA-Compatible" content="IE=edge">
-    <meta name="viewport" content="width=device-width,initial-scale=1">
-    <meta name="dеѕсrірtіоn" content="%APP_DESCRIPTION%">
-    <link id="favicon" rel="shortcut icon" href="%APP_FAVICON%">
-    <link id="manifest" rel="manifest" href="%APP_MANIFEST%">
-    <link href="/css/front.css" rel="stylesheet">
-</head>
-<body>
-<noscript>You need to enable JavaScript to run this app.</noscript>
-<div id="root"></div>
-</body>
-</html>
