--- opencti-platform/opencti-front/builder/plugin/esbuild-relay.js
@@ -1,35 +1,0 @@
-const {promises} = require('fs');
-const crypto = require('crypto');
-const {print, parse} = require('graphql');
-
-module.exports.RelayPlugin = {
-    name: 'relay',
-    setup: (build) => {
-        build.onLoad({filter: /\.(ts|tsx|js|jsx)$/, namespace: "file"}, async (args) => {
-            let contents;
-            if (args.path.includes('src') && !args.path.includes('node_modules') && !args.path.includes('__generated__')) {
-                contents = await promises.readFile(args.path, 'utf8');
-                if (contents.includes('graphql`')) {
-                    const imports = [];
-                    contents = contents.replaceAll(/\sgraphql`([\s\S]*?)`/gm, (match, query) => {
-                        const formatted = print(parse(query));
-                        const graphLineExtra = formatted.split("\n").map(() => "\n").join(""); // Need for debug alignment
-                        const name = /(fragment|mutation|query|subscription) (\w+)/.exec(formatted)[2];
-                        const id = `graphql__${crypto.randomBytes(10).toString('hex')}`;
-                        imports.push(`import ${id} from "./__generated__/${name}.graphql";`);
-                        return id + graphLineExtra;
-                    });
-                    contents = imports.join('\n') + contents;
-                }
-            }
-            let loader = 'tsx';
-            if (args.path.endsWith('.js')) loader = 'js';
-            if (args.path.endsWith('.jsx')) loader = 'jsx';
-            if (args.path.endsWith('.ts')) loader = 'ts';
-            return {
-                contents,
-                loader,
-            };
-        });
-    },
-};
