+++ opencti-platform/opencti-front/package.json
@@ -1,6 +1,7 @@
 {
   "name": "opencti-front",
   "version": "7.260417.0",
+  "type": "module",
   "private": true,
   "workspaces": [
     "packages/*"
@@ -9,11 +10,10 @@
     "relay": "relay-compiler relay.config.json",
     "dev": "yarn relay && vite",
     "dev:vite:debug": "yarn relay && vite --debug plugin-transform",
-    "start": "yarn relay && node builder/dev/dev.js",
+    "start": "yarn dev",
     "esbuild": "node builder/prod/prod.js",
-    "build": "yarn relay && yarn esbuild",
-    "build:standalone": "yarn relay && yarn esbuild --keep",
-    "lint": "DEBUG=eslint:cli-engine TIMING=1 eslint --quiet --cache --config eslint.config.mjs src",
+    "build": "yarn relay && vite build",
+    "lint": "DEBUG=eslint:cli-engine TIMING=1 eslint --quiet --cache --config eslint.config.js src",
     "control": "yarn audit --groups dependencies --summary",
     "test": "vitest run",
     "test:watch": "vitest watch",
@@ -104,7 +104,6 @@
     "formik-mui": "5.0.0-alpha.0",
     "formik-mui-lab": "1.0.0",
     "graphiql": "5.2.2",
-    "graphql": "16.13.2",
     "graphql-ws": "6.0.8",
     "html-react-parser": "6.0.1",
     "html-to-image": "1.11.13",
@@ -190,18 +189,13 @@
     "@vitejs/plugin-react": "6.0.1",
     "@vitest/coverage-v8": "4.1.4",
     "babel-plugin-relay": "20.1.1",
-    "chokidar": "5.0.0",
-    "compression": "1.8.1",
-    "esbuild": "0.28.0",
     "eslint": "10.2.0",
     "eslint-import-resolver-typescript": "4.4.4",
     "eslint-plugin-custom-rules": "workspace:*",
     "eslint-plugin-import": "2.32.0",
     "eslint-plugin-import-newlines": "2.0.0",
     "eslint-plugin-react": "7.37.5",
-    "express": "5.2.1",
     "globals": "17.4.0",
-    "http-proxy-middleware": "3.0.5",
     "i18n-auto-translation": "2.2.3",
     "jsdom": "29.0.2",
     "license-checker-rseidelsohn": "4.4.2",
