+++ opencti-platform/opencti-front/packages/eslint-plugin-custom-rules/package.json
@@ -1,11 +1,9 @@
 {
   "name": "eslint-plugin-custom-rules",
   "private": true,
+  "type": "module",
   "main": "lib/index.js",
   "exports": "./lib/index.js",
-  "dependencies": {
-    "requireindex": "1.2.0"
-  },
   "peerDependencies": {
     "eslint": ">=7"
   }
