+++ opencti-platform/opencti-front/script/extract-label-keyword.js
@@ -1,11 +1,5 @@
-const fs = require('fs');
-const util = require('util');
-const path = require('path');
-
-const readdir = util.promisify(fs.readdir);
-const stat = util.promisify(fs.stat);
-const readFile = util.promisify(fs.readFile);
-const writeFile = util.promisify(fs.writeFile);
+import { readdir, stat, readFile, writeFile } from 'node:fs/promises';
+import path from 'node:path';
 
 const srcDirectory = '../opencti-graphql/src';
 const englishTranslationFiles = 'lang/back/en.json';
@@ -16,12 +10,12 @@ const extractedValues = {};
 // extract all the 'label' of attributes / relation refs from the backend schema
 // and add them in opencti-front/lang/en-back.json
 
-function extractValueFromPattern(pattern) {
+const extractValueFromPattern = (pattern) => {
   const match = /label: '([^']+)'/.exec(pattern);
   return match ? match[1] : null;
 }
 
-async function extractI18nValues(directory) {
+const extractI18nValues = async (directory) => {
   try {
     const files = await readdir(directory);
     for (const file of files) {
@@ -49,7 +43,7 @@ async function extractI18nValues(directory) {
   }
 }
 
-async function mergeWithExistingData() {
+const mergeWithExistingData = async () => {
   try {
     const existingData = await readFile(englishTranslationFiles, 'utf8');
     const existingValues = JSON.parse(existingData);
@@ -79,10 +73,6 @@ async function mergeWithExistingData() {
   }
 }
 
-async function main() {
-  console.log('--- extract labels from backend schema definition ---');
-  await extractI18nValues(srcDirectory);
-  await mergeWithExistingData();
-}
-
-main();
+console.log('--- extract labels from backend schema definition ---');
+await extractI18nValues(srcDirectory);
+await mergeWithExistingData();
