+++ opencti-platform/opencti-front/script/sort-translation-files.js
@@ -1,19 +1,19 @@
-const fs = require('fs');
-const path = require('path');
+import { readFile, readdir } from 'node:fs/promises';
+import path from 'node:path';
 
-function sortJSONKeys(json) {
+const sortJSONKeys = (json) => {
   const sortedKeys = Object.keys(json).sort((a, b) => a.toLowerCase().localeCompare(b.toLowerCase()));
   const sortedJSON = {};
-  sortedKeys.forEach(key => {
+  sortedKeys.forEach((key) => {
     sortedJSON[key] = json[key];
   });
   return sortedJSON;
 }
 
-function sortJSONFile(filePath) {
+const sortJSONFile = async (filePath) => {
   try {
     // Read JSON file
-    const jsonData = JSON.parse(fs.readFileSync(filePath, 'utf8'));
+    const jsonData = JSON.parse(await readFile(filePath, 'utf8'));
     
     // Sort JSON keys
     const sortedJSON = sortJSONKeys(jsonData);
@@ -27,13 +27,13 @@ function sortJSONFile(filePath) {
   }
 }
 
-function sortAllJSONFiles(dirPath) {
+const sortAllJSONFiles = async (dirPath) => {
   try {
     // Get list of files in directory
-    const files = fs.readdirSync(dirPath);
+    const files = await readdir(dirPath);
     
     // Iterate through files
-    files.forEach(file => {
+    files.forEach((file) => {
       const filePath = path.join(dirPath, file);
       
       // Skip if not a JSON file or if it's en.json
@@ -49,5 +49,5 @@ function sortAllJSONFiles(dirPath) {
   }
 }
 
-sortAllJSONFiles('lang/front');
-sortAllJSONFiles('lang/back');
+await sortAllJSONFiles('lang/front');
+await sortAllJSONFiles('lang/back');
